refactor: extract utility functions to src/main/utils.ts
